requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
patsy==0.5.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scipy==1.14.1
six==1.16.0
statsmodels==0.14.4
tzdata==2024.2
Werkzeug==3.0.6
